pyproject.toml
[tool.poetry]
name = "quickscope-interview-bot"
version = "0.1.0"
description = "LangGraph-powered current-state process interview bot"
authors = ["QuickScope Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
quickscope = "src.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.2.0"
langgraph-cli = "^0.1.0"
langchain-core = "^0.3.0"
langchain-openai = "^0.2.0"
pydantic = "^2.0"
pydantic-settings = "^2.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
